[#13] refactor: split DocumentWalker/HtmlRenderer into cpp files
