+++ HwpxToMarkdown.vcxproj.filters
@@ -24,6 +24,12 @@
     <ClCompile Include="src\sdk\SDK_Wrapper.cpp">
       <Filter>소스 파일</Filter>
     </ClCompile>
+    <ClCompile Include="src\walker\DocumentWalker.cpp">
+      <Filter>소스 파일</Filter>
+    </ClCompile>
+    <ClCompile Include="src\render\HtmlRenderer.cpp">
+      <Filter>소스 파일</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <Library Include="lib\Owpml.lib" />
